Fix ucrt source path when packaging the Windows toolchain
Recent versions of the 10.0.17763 SDK have rearranged the redist
directory. This change adjusts the packaging so that we can package the
SDK with and without this change. The script looks first in the
versioned directory and if that doesn't exist then it falls back to the
unversioned directory.

Similar changes will be needed in the Chromium build\vs_toolchain.py
script.

Change-Id: I835bf45f613a0aebb56eb1e8e63e6ffa5252edf0