Avoid really slow OWNERS operations for large/huge patches
If many different OWNERS files or "set noparent" rules are in scope
of a patch, the owners system could become slow since it traversed and
ran fnmatch for all the files and all the rules for every file.

By splitting the rules by longest glob-free directory path, you get
much smaller lists to process, which makes the code much faster.

(Test case with "all the files" went from >15 minutes to 30 seconds).

The addition of many noparent rules in the Chromium tree might have been
the major reason this is needed.

Bug: 920591
Change-Id: I59590f61b2345e2d54073e68e1792fa92d10cd90