Update patch set 3
Patch Set 3:

Thanks for jumping on this! It's been on my list of things I'm worried about for a while now. I'm not that familiar with these systems so I'd prefer to leave the LG to others, but love the change.

Patch-set: 3
Reviewer: Gerrit User 1303432 <1303432@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>