Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > To be clear: I'll wait for you to tell me whether you want me to change this to be the default or whether I should land as-is before landing the CL.
> 
> I was mostly asking Dirk if this should be the default :)

I think it probably shouldn't be the default, at least not for now. There are cases where you might be an owner for part of the files in the CL, but need owners for other files that you're modifying, and I wouldn't want to have to know to filter out the stuff I own.

I have no idea how common that case is compared the case where you are using this to find a reviewer for stuff you own. 

@ehmaldonaldo, is this something the depot_tools metrics might be able to tell us, or do we not have that level of detail in them?

Patch-set: 2
CC: Gerrit User 1167909 <1167909@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Reviewer: Gerrit User 1003232 <1003232@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>