Check for impossible date ranges
If you specify the wrong year to the -b flag then you may end up with an
inverted time range that is entirely in the future, and this causes odd
failures. Or at least that's what I've heard. This detects some of these
error cases and halts.

Also a one-line git cl format cleanup.

Change-Id: Iede80faed00d4857443b3a1d853fa2ba69f47023