Update patch set 2
Patch Set 2:

> Patch Set 1:
> 
> If this doesn't do anything than apply window dressing to my fix, I'd prefer to not do this. There's a comment explaining why the import is done on-demand right above where you're adding the import. Most imports in gclient_utils.py probably also take "no time" to import, and still this change made the gn and clang-format wrappers twice as fast somehow.

I've updated the comment.

I'm not disagreeing with the basic idea that we should be careful about what gets imported in these wrappers, but we have to balance that against the idea that importing modules just-in-time can be hard to reason about and lead to bugs like the one you introduced, which is why it's usually considered bad style.

For reference, importing io seems to take ~600 usecs on my linux box, and logging ~3100 usecs. I'd argue that we should probably import io at the top-level if the uses can't be eliminated, and remove the call to logging since I'm not sure that it's even really useful ...

Patch-set: 2
Reviewer: Gerrit User 1003232 <1003232@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>