Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 1:
> > 
> > If this doesn't do anything than apply window dressing to my fix, I'd prefer to not do this. There's a comment explaining why the import is done on-demand right above where you're adding the import. Most imports in gclient_utils.py probably also take "no time" to import, and still this change made the gn and clang-format wrappers twice as fast somehow.
> 
> I've updated the comment.
> 
> I'm not disagreeing with the basic idea that we should be careful about what gets imported in these wrappers, but we have to balance that against the idea that importing modules just-in-time can be hard to reason about and lead to bugs like the one you introduced, which is why it's usually considered bad style.
> 
> For reference, importing io seems to take ~600 usecs on my linux box, and logging ~3100 usecs. I'd argue that we should probably import io at the top-level if the uses can't be eliminated, and remove the call to logging since I'm not sure that it's even really useful ...

I agree with removing the call to logging :)

io is only used in a single place, so there's no advantage to keeping the import where it is.

I'll point out that the bug I introduced was due to me not importing sys at all, which mostly points to missing test coverage for these functions. The targeted imports were added in the fix, and were correct. So I don't think based on the history of this file there's any data to say that the local import style is buggier. (In general I'm willing to believe that, but it also leads to modules that load slower. And the code in this new gclient_paths module hasn't changed in many years and is used in several places that are thin wrappers around real executables, so a local important does seem like the right tradeoff to me.)

Patch-set: 2
Reviewer: Gerrit User 1001629 <1001629@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>