Avoiding writing ninjalog.cfg unnecessarily
Writing to ninjalog.cfg comes with the risk of failure (Ctrl+C at the
wrong time) which can lead to an empty or corrupted file. This will
cause errors on subsequent builds when loading the config file, and
will lose the users opt-in/out settings.

This change ensures that the config file is only written when something
has changed, thus essentially avoiding the risk.

Bug: 941810
Change-Id: If266f07536569c9ac91ee37c28797c8b2ccc6fdc