Update patch set 1
Patch Set 1:

(1 comment)

The current behavior of doing the decrementing in LoadConfig appears to be causing confusion. Should I clean that up at the same time so that the behavior is clearer?

As in, explicitly decrement the count if it is greater than zero and then save?

Patch-set: 1
Reviewer: Gerrit User 1128439 <1128439@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0
Work-in-progress: false