Update patch set 4
Patch Set 4:

For a longer term thing; we can set the contents of this file when building the git cipd package too (so then it's 'already set' once you unpack the cipd package). Once we verify that protocol 2 works everywhere, we should probably spin a new git version with it set.

Patch-set: 4
CC: Gerrit User 1001907 <1001907@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>