Update patch set 1
Patch Set 1: Code-Review+1

(1 comment)

Sorry for the long delay on this. 

After mulling it over, I think the feature makes sense, i.e., there should be a way to keep re-warning people.

I wonder a little bit about what the right default is, i.e., I can imagine a choice of 'always', 'once', 'never' (where today we do 'once', and this CL changes it to 'always'). I kinda wonder if we should have an actual flag for this, but I think I'm inclined to try and land this like it is, and only add a flag if there's demand for it.

+mmoss, +agable, in case they have thoughts on this. Please wait a bit to give them a chance to comment before landing.

Patch-set: 1
CC: Gerrit User 1002222 <1002222@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
CC: Gerrit User 1002100 <1002100@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Reviewer: Gerrit User 1003232 <1003232@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1