Update patch set 4
Patch Set 4: Auto-Submit+1

> Patch Set 4: Code-Review+1
> 
> (2 comments)
> 
> Sorry for the long delay in reviewing this. It basically LGTM.
> 
> That said, I wrote nearly all of this 6+ years ago, and looking at it now, I really wonder about my choices of algorithms and data structures. I think your choices look like reasonable extensions, but life might be even better with a more extensive rewrite at some point.

I suspect it's that the world has changed since the code was written and that the code was plenty good when it was new and shiny. I'm of the school that "simple is good" so if simple worked, then why not.

I updated the comments a bit. 

Do we have any good testing of this? I used the code a lot when extracting owner statistics so I know it returns about the same as the previous code, but maybe there is some edge case missed somewhere.

Patch-set: 4
Reviewer: Gerrit User 1114374 <1114374@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Auto-Submit=+1