Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> The tests pass for me locally, so I'm not sure how to update the test. I guess it means that we'll have to land this roll together with https://chromium-review.googlesource.com/c/chromium/tools/build/+/1570229 ?

It's not a failing test. It's a warning that your changes will affect expectations in the downstream build repo - see text next to "result" step in the failing build. You'll have to add the footer and then expect a recipe roll to be created (usually within a few hours, looking similar to https://crrev.com/c/1564980). The reason this is required is to make sure that you don't forget to review and LGTM that roll. Without your approval the roller will get stuck and other people won't be able to see their changes rolled downstream. You can also review the expectation changes in the logs of the failed steps, but the diff is much harder to read than in Gerrit, so I rarely do that.

Patch-set: 2
Reviewer: Gerrit User 1153089 <1153089@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>