Update patch set 1
Patch Set 1:

+agable for his potential thoughts as well as iannucci's

commenting here, though I could've equally said this on the bug:

I don't fully understand the motivation for this change. 

If this was for chromium specifically (which was the example in the bug), we don't support submodules and wouldn't want you to add them, so it seems correct to me that we'd report an error rather than ignoring them.

If this *wasn't* for chromium specifically (which is what the CL description suggests, and in which case I'm guessing you used chromium in the bug description just as an example), I could maybe see that this might be good behavior, but I'm not yet sure.

We do want to move to supporting submodules better, but I don't know how we want to do that. I'm leery of a change that would change the behavior globally (across all projects) like this, and you might want this to be something that you'd either opt in to, or at least be able to opt out of.

+iannucci, agable, thoughts?

Patch-set: 1
Reviewer: Gerrit User 1002100 <1002100@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Reviewer: Gerrit User 1003232 <1003232@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>