Update patch set 8
Patch Set 8:

The issue is that applying patches to branch heads doesn't work (at least when --with-branch-heads is not specified):
https://ci.chromium.org/p/webrtc/builders/try/linux_compile_rel/13395

This is because gclient_scm doesn't know how to handle the target_branch for the patch.

This change should fix that:
- chromium/src patch on top of refs/heads/master: https://luci-milo.appspot.com/swarming/task/44955af05d321b10
- widevine change on top of an experimental branch: https://luci-milo.appspot.com/swarming/task/4495643d831c3710?server=chrome-swarming.appspot.com
- webrtc change on top of refs/branch-heads/m75

The most common case is refs/heads/master, for which the target branch should be refs/remotes/origin/master.

The second most common case is when we're patching the top solution, and the target branch is the branch the solution was synced to.
In this case, bot_update.py fetches the necessary refs, so we don't have to do anything:
https://cs.chromium.org/chromium/tools/depot_tools/recipes/recipe_modules/bot_update/resources/bot_update.py?l=721

In any other case, we check if the target branch we were given (e.g. refs/heads/foo) is already available, or if the remote for that branch (e.g. refs/remotes/origin/foo) is available.
Otherwise we fetch the remote ref.
See https://luci-milo.appspot.com/swarming/task/4495643d831c3710?server=chrome-swarming.appspot.com for a sample run of this (this is the widevine task from before I figured the scm.GIT.IsValidRevision bug on windows)

So we shouldn't be making any more fetches than necessary.

Patch-set: 8
Reviewer: Gerrit User 1167909 <1167909@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>