Update patch set 1
Patch Set 1:

(2 comments)

I vaguely remember discussing whether to support 'in' with phajdan when we were spec'ing and implementing this. I think we ended up not implementing it because we weren't sure we needed it, but I don't see anything in the spec (go/gclient-conditionals) or remember anything that says that something like this was a bad idea.

So, I think the idea is okay, but I think you need to address the comments I left, and you need to also add some tests (see tests/gclient_eval_unittest.py).

We should also update the spec after this lands and sticks.

Patch-set: 1
Reviewer: Gerrit User 1003232 <1003232@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>