Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1: Code-Review-1
> > 
> > It looks like fetch.py is already expected to be python3-compatible, e.g.:
> > https://chromium.googlesource.com/chromium/tools/depot_tools/+/c2f74c1698a8e75f9474b92dcd69c73178d240b8
> > 
> > so rather than forcing python2 here, which will probably cause its own problems, the right thing is to probably fix whatever is breaking with python3. Do you have any details or bug describing the failures you're seeing? Maybe add something to crbug.com/939847
> 
> But not working here, it's python2 compatible all of the script is in python2, I have to force my system to change the default python version from 3 to 2 to run the script and build.

But you shouldn't have to do that. It should work with the default python3 install. If it's not, that's a bug in fetch.py and needs to be fixed. Please add details of the python3 errors you're encountering to crbug.com/939847.

Patch-set: 1
Reviewer: Gerrit User 1002222 <1002222@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>