Update patch set 1
Patch Set 1:

Hi! I do know this code. One question before I dive in: Do you have a bug or doc tracking this work and the motivation behind it?

I ask because we have a long history of not writing PRESUBMIT checks which modify the state of the world. This is particularly important because PRESUBMIT runs during "git cl upload", at which point you need to already have a clean git commit to be uploaded -- modifying files and generating dirty state during upload is not something we believe devs want.

Patch-set: 1