gclient_utils: buffer Unicode characters for decoding in CheckCallAndFilter
Python 3 requires writes to stdout to be strings.
Python 2 doesn't care.

The output from the child process is bytes, so they simply need to be decoded to strings and then output.
However, since the code buffers byte by byte, this doesn't work for Unicode characters, because they're multiple bytes, so decoding a single character will fail.

The solution is to buffer the bytes and try to decode when outputting.
If successful, simply clear the buffer.
If not, don't clear the buffer, add another byte and try again on the next iteration.

There was an earlier attempt to fix this issue in chromium:1524583, which didn't do any buffering, breaking Unicode output.
That change was reverted and a test added for Unicode output in chromium:1541334.

Bug: 942522
Change-Id: I9ad97c739eca201fc5558ba3c9b87c041bd807a4