Update patch set 2
Patch Set 2: Code-Review+1

As discussed in person, it seems a little weird that passing this flag modifies the CL description (which effectively will make it "sticky" in the sense that all subsequent patch sets will also not be cancelled, unless someone goes and removes the footer).

You could have the upload remove the footer if the flag isn't passed, but that might be overkill.

We need to have the footer one way or another to tell the CQ what to do (and provide a visible audit trail).

I think it's okay to leave as-is. lgtm.

Patch-set: 2
Reviewer: Gerrit User 1003232 <1003232@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1