Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Bump; is there any initial set of comments that I could start working on while this is being thoroughly reviewed?

Hi! Sorry for the reply. I can also review this, but I don't fully understand your response to agable@, and it seems like there's a lot of different stuff in this CL.

Is it possible to break this up into multiple pieces? E.g., if you have something that needs to handle a different branching/upload model, can we separate that out from changing the presubmit api?

Also, if the devtools presubmit check is actually modifying the files, that's just flat-out a bug that we should fix, it's not okay for them to do this.

Patch-set: 1