Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > Bump; is there any initial set of comments that I could start working on while this is being thoroughly reviewed?
> 
> Hi! Sorry for the reply. I can also review this, but I don't fully understand your response to agable@, and it seems like there's a lot of different stuff in this CL.
> 
> Is it possible to break this up into multiple pieces? E.g., if you have something that needs to handle a different branching/upload model, can we separate that out from changing the presubmit api?
> 
> Also, if the devtools presubmit check is actually modifying the files, that's just flat-out a bug that we should fix, it's not okay for them to do this.

I'll try to split it up into multiple changes. The crux of the issue, here, is that if you provide a `base_branch` to `git cl presubmit` (as a positional parameter), the determinations from that (what the branches for the diff are, etc) don't properly get passed through to reentrant calls into the depot_tools machinery. This caused issues for us where some issues weren't reproducing on dev machines after hitting in CI - because the dev machines had pushed branches, setting git's `upstream` value for the branch, while in CI this hadn't been set, so it defaulted to `origin/master`.

The only change to the presubmit api, in this change, is that calling the formatter is explicitly part of the api, not implicitly (by virtue of being available in PATH). By making this a presubmit api explicitly, we can pass through the Change object directly, avoiding any issues around redetermining the diff.

The other changes are mostly basic code-cleanliness and bugfixing - the format code was just one long function, and there were a couple places where a sufficiently large diff causes it to error out. I can pull these fixes out to separate CLs.

Patch-set: 1
Reviewer: Gerrit User 1355758 <1355758@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0