Reland "depot_tools: Make gclient_utils Python 3 compatible."
This is a reland of 602076dcf209cf2630e3976c648f4f4d8da0b553

The original change failed on input 0xe2 0x9c 0x94 which is a utf-8 heavy check mark.
The problem is that CheckCallAndFilter was processing output byte by byte,
so trying to decode only one of utf-8 bytes in sequence will always fail.

This wasn't catched by the test because the process stdout was mocked to be a
StringIO, when it should have been a BytesIO.

The test and the code should be fixed now.

Original change's description:
> depot_tools: Make gclient_utils Python 3 compatible.
>
> Bug: 984182
> Change-Id: Ifb93466dcb7541f46c31c9b298049c425dd150d6
> Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/depot_tools/+/1715472
> Reviewed-by: Andrii Shyshkalov <tandrii@chromium.org>
> Commit-Queue: Edward Lesmes <ehmaldonado@chromium.org>

Bug: 984182
Change-Id: Ida30e295b872c8c1a1474a376a90aecea924f307