pylint: rework wrapper to use vpython for multiple versions
This allows people to pin a specific version of pylint when running
it.  The default behavior is unchanged: we run the pylint copy that
is in depot_tools now (1.5.6).  We also add in support for selecting
the 1.8 series as an example of how vpython integration would work.

The plan is to make pylint-1.5 available via vpython too so we can
drop the copy from depot_tools entirely.

Bug: 866772
Test: `./pylint ...` still works
Test: `./pylint --pylint-version=1.5 ...` selects the local 1.5 copy
Test: `./pylint --pylint-version=1.8 ...` selects the vpython 1.8 copy
Change-Id: I7c47187493564db81e3eb28c6dbd09e6309ead46