Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> What would be the advantage of doing this? It's more consistent if autoninja always supplies the value, and it makes it visible.
> 
> Also, ninja.exe is buggy on machines with > 64 processors, whereas autoninja handles them correctly.

I guess I was assuming that upstream ninja developers know what they are doing here and that we don't want to duplicate the logic of choosing a good number of local jobs to kick of.

Another way of putting it:  My expectation was that for non-goma builds autoninja was basically a no-op so I was surprised to see it injecting a -j flag.

If upstream ninja is buggy then maybe we should try to get that fixed and have a TODO here that says why we are not trusting it?

Patch-set: 1
Reviewer: Gerrit User 1002666 <1002666@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0