Update patch set 1
Patch Set 1:

> I guess I was assuming that upstream ninja developers know what they are doing here and that we don't want to duplicate the logic of choosing a good number of local jobs to kick of.
> 
> Another way of putting it:  My expectation was that for non-goma builds autoninja was basically a no-op so I was surprised to see it injecting a -j flag.
> 
> If upstream ninja is buggy then maybe we should try to get that fixed and have a TODO here that says why we are not trusting it?

When the "always-add -j" logic was added we didn't know that ninja.exe was buggy. The fix for that was added to autoninja.py a few weeks ago by jessemckenna@ (CCed). Adding a comment isn't a bad idea, but it is not universally agreed that always passing -j is a bad idea. It was added intentionally because it was felt to be more consistent. You'd have to talk to gab@ about that.

Patch-set: 1
Reviewer: Gerrit User 1128439 <1128439@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>