setup_color: Don't output an error if GetConsoleMode fails
Turns out GetConsoleMode fails if the user's console isn't native. This is sometimes the case for Git Bash.

Also moved the checking code out of the if block so it applies in generic console cases too.
This went unnoticed, because setup_color.py is currently only really used in Git scripts, which are piped.

Change-Id: I93357e479f84122f759e419d15c02500809657d3