Update patch set 1
Patch Set 1:

How does this look?
I could also add checking for the 'false' options and error-out if it is set to an illegal value, but I'm not sure it is worth doing that given that goma will be doing that checking. This just fixes a misleading comment and ensures that use_goma is set to False exactly when it should be.

Patch-set: 1
Reviewer: Gerrit User 1161379 <1161379@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Reviewer: Gerrit User 1128439 <1128439@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Work-in-progress: false